Move FollowRedirects to Middleware::Response
This moves the FollowRedirects middleware to
Octokit::Middleware::Response to better indicate its utility as Faraday
middleware.
