+++ lib/octokit/default.rb
@@ -1,4 +1,4 @@
-require 'octokit/middleware/follow_redirects'
+require 'octokit/middleware/response/follow_redirects'
 require 'octokit/response/raise_error'
 require 'octokit/response/feed_parser'
 require 'octokit/version'
@@ -25,7 +25,7 @@ module Octokit
 
     # Default Faraday middleware stack
     MIDDLEWARE = RACK_BUILDER_CLASS.new do |builder|
-      builder.use Octokit::Middleware::FollowRedirects
+      builder.use Octokit::Middleware::Response::FollowRedirects
       builder.use Octokit::Response::RaiseError
       builder.use Octokit::Response::FeedParser
       builder.adapter Faraday.default_adapter
